Move CA2002 to its new location - no functional change. Delete fixer stub since we cannot offer a fix for this rule.
