+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.props
@@ -13,6 +13,8 @@
         -Microsoft.Design#CA1003;
         -Microsoft.Design#CA1009;
         -Microsoft.Design#CA1019;
+        
+        -Microsoft.Reliability#CA2002;
 
         -Microsoft.Globalization#CA1309;
       </CodeAnalysisRuleSetOverrides>
