+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.props
@@ -14,6 +14,8 @@
         -Microsoft.Design#CA1009;
         -Microsoft.Design#CA1019;
 
+        -Microsoft.Reliability#CA2002;
+
         -Microsoft.Globalization#CA1309;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
