--- PortFxCop/src/System.Runtime.Analyzers/CSharp/CSharpDoNotLockOnObjectsWithWeakIdentity.Fixer.cs
@@ -1,20 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-                                 
-using System.Composition;
-using System.Diagnostics;  
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace System.Runtime.Analyzers
-{                                 
-    /// <summary>
-    /// CA2002: Do not lock on objects with weak identity
-    /// </summary>
-    [ExportCodeFixProvider(LanguageNames.CSharp), Shared]
-    public class CSharpDoNotLockOnObjectsWithWeakIdentityFixer : DoNotLockOnObjectsWithWeakIdentityFixer
-    { 
-
-    }
-}
